requirements.txt
python-decouple
requests
sqlalchemy
pandas
beautifulsoup4